# Add missing optimization for `iterate(::LogicalIndex{<:CartesianIndex,<:BitArray})`