@N5N3 2021-12-18
Profile shows that the slow speed for `IndexCartesian` comes from the unoptimized fallback.
So I add the missing optimization with `findall` as a reference.
This will also make `A[A.>0] = B` faster.