@N5N3 2021-12-20
Update the Benchmark
```julia
using Random, BenchmarkTools
Random.seed!(0)
a = randn(128, 32);
A = view(a, axes(a)...);
A2 = view(reshape(a,32,32,4),1:32,1:32,1:4);
p = a .> 0;
@btime view($a, $p); # 1.630 μs  -->  1.360 μs
@btime view($A, $p); # 16.000 μs --> 2.500 μs
@btime view($A2, reshape($p, size($A2))); # 13.100 μs --> 3.638 μs
```