@N5N3 2022-02-16
We already have
https://github.com/JuliaLang/julia/blob/f09099206c7d11ffbe25e3416d504dcbdd3de7ec/test/abstractarray.jl#L458-L462

So the added optimization should have been tested.